clean up documentation and imports
